feat: lifecycle events for scratch org create
