+++ src/exported.ts
@@ -104,7 +104,8 @@ export { PermissionSetAssignment, PermissionSetAssignmentFields } from './org/pe
 
 export { ScratchOrgCreateOptions, ScratchOrgCreateResult, scratchOrgCreate } from './org/scratchOrgCreate';
 
-export { ScratchOrgInfo } from './org/scratchOrgInfoApi';
+export { ScratchOrgInfo } from './org/scratchOrgTypes';
+export { ScratchOrgLifecycleEvent, scratchOrgLifecycleEventName } from './org/scratchOrgLifeCycleEvents';
 
 // Utility sub-modules
 export * from './util/sfdc';
