+++ src/org/scratchOrgTypes.ts
@@ -1,0 +1,50 @@
+/*
+ * Copyright (c) 2020, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+import { JsonMap } from '@salesforce/ts-types';
+
+export interface ScratchOrgInfo {
+  AdminEmail?: string;
+  readonly CreatedDate?: string;
+  ConnectedAppCallbackUrl?: string;
+  ConnectedAppConsumerKey?: string;
+  Country?: string;
+  Description?: string;
+  DurationDays?: string;
+  Edition?: string;
+  readonly ErrorCode?: string;
+  readonly ExpirationDate?: string;
+  Features?: string;
+  HasSampleData?: boolean;
+  readonly Id?: string;
+  Language?: string;
+  LoginUrl: string;
+  readonly Name?: string;
+  Namespace?: string;
+  OrgName?: string;
+  Release?: 'Current' | 'Previous' | 'Preview';
+  readonly ScratchOrg?: string;
+  SourceOrg?: string;
+  readonly AuthCode: string;
+  Snapshot: string;
+  readonly Status: 'New' | 'Creating' | 'Active' | 'Error' | 'Deleted';
+  readonly SignupEmail: string;
+  readonly SignupUsername: string;
+  readonly SignupInstance: string;
+  Username: string;
+  settings?: Record<string, unknown>;
+  objectSettings?: { [objectName: string]: ObjectSetting };
+  orgPreferences?: {
+    enabled: string[];
+    disabled: string[];
+  };
+}
+
+export interface ObjectSetting extends JsonMap {
+  sharingModel?: string;
+  defaultRecordType?: string;
+}
